+++ configurate-core/src/main/java/ninja/leaping/configurate/Types.java
@@ -23,7 +23,8 @@ import java.text.ParseException;
 import java.time.DateTimeException;
 import java.time.Instant;
 import java.time.format.DateTimeParseException;
-import java.time.temporal.*;
+import java.time.temporal.ChronoField;
+import java.time.temporal.TemporalAccessor;
 import java.util.Date;
 
 /**
@@ -341,11 +342,34 @@ public final class Types {
         return value instanceof Boolean ? (Boolean) value : null;
     }
 
+    /**
+     * Attempts to convert <code>value</code> to a {@link Date}.
+     *
+     * <p>
+     * <ul>
+     *     <li>If <code>value</code> is a {@link Date}, casts and returns</li>
+     *     <li>If <code>value</code> is a {@link TemporalAccessor}, returns a converted value</li>
+     *     <li>If <code>value</code> is a {@link Number}, returns a converted as if the number is a
+     *     unix timestamp in milliseconds</li>
+     *     <li>Otherwise, <code>value</code> is converted {@link Object#toString() to a string}, and
+     *     then {@link DateFormat#parse(String) parsed}.</li>
+     * </ul>
+     *
+     * <p>Returns null if <code>value</code> is null.</p>
+     *
+     * @param value The value
+     * @return <code>value</code> as a {@link Date}, or null
+     */
+    @Nullable
     public static Date asDate(Object value) {
-        if (value == null) return null;
+        if (value == null) {
+            return null;
+        }
+
         if (value instanceof Date) {
             return (Date) value;
         }
+
         if (value instanceof TemporalAccessor) {
             TemporalAccessor temp = (TemporalAccessor) value;
             if (temp.isSupported(ChronoField.INSTANT_SECONDS)) {
@@ -356,9 +380,11 @@ public final class Types {
                 return new Date(millis);
             }
         }
+
         if (value instanceof Number) {
             return new Date(((Number) value).longValue());
         }
+
         String dateString = value.toString();
         try {
             return DateFormat.getInstance().parse(dateString);
@@ -367,6 +393,13 @@ public final class Types {
         }
     }
 
+    /**
+     * Returns <code>value</code> if it is a {@link Date}.
+     *
+     * @param value The value
+     * @return <code>value</code> as a {@link Date}, or null
+     */
+    @Nullable
     public static Date strictAsDate(Object value) {
         if (value == null) {
             return null;
@@ -375,14 +408,39 @@ public final class Types {
         return value instanceof Date ? (Date) value : null;
     }
 
+    /**
+     * Attempts to convert <code>value</code> to a {@link Instant}.
+     *
+     * <p>
+     * <ul>
+     *     <li>If <code>value</code> is a {@link Instant}, casts and returns</li>
+     *     <li>If <code>value</code> is a {@link Date}, returns a converted value</li>
+     *     <li>If <code>value</code> is a {@link TemporalAccessor}, returns a converted value</li>
+     *     <li>If <code>value</code> is a {@link Number}, returns a converted as if the number is a
+     *     unix timestamp in milliseconds</li>
+     *     <li>Otherwise, <code>value</code> is converted {@link Object#toString() to a string}, and
+     *     then {@link Instant#parse(CharSequence) parsed}.</li>
+     * </ul>
+     *
+     * <p>Returns null if <code>value</code> is null.</p>
+     *
+     * @param value The value
+     * @return <code>value</code> as a {@link Instant}, or null
+     */
+    @Nullable
     public static Instant asInstant(Object value) {
-        if (value == null) return null;
+        if (value == null) {
+            return null;
+        }
+
         if (value instanceof Instant) {
             return (Instant) value;
         }
+
         if (value instanceof Date) {
             return Instant.ofEpochMilli(((Date) value).getTime());
         }
+
         if (value instanceof TemporalAccessor) {
             try {
                 return Instant.from((TemporalAccessor) value);
@@ -390,6 +448,7 @@ public final class Types {
                 return null;
             }
         }
+
         String dateString = value.toString();
         try {
             return Instant.parse(dateString);
@@ -398,6 +457,13 @@ public final class Types {
         }
     }
 
+    /**
+     * Returns <code>value</code> if it is a {@link Instant}.
+     *
+     * @param value The value
+     * @return <code>value</code> as a {@link Instant}, or null
+     */
+    @Nullable
     public static Instant strictAsInstant(Object value) {
         if (value == null) {
             return null;